Refactor model transfer jobs handler to use downstream repository layer
